+++ pyccel/ast/builtins.py
@@ -20,6 +20,7 @@ from .datatypes import GenericType, PythonNativeComplex, PrimitiveComplexType
 from .datatypes import HomogeneousTupleType, InhomogeneousTupleType
 from .datatypes import HomogeneousListType, HomogeneousContainerType
 from .datatypes import FixedSizeNumericType, HomogeneousSetType, SymbolicType
+from .datatypes import DictType
 from .internals import PyccelFunction, Slice, PyccelArrayShapeElement
 from .literals  import LiteralInteger, LiteralFloat, LiteralComplex, Nil
 from .literals  import Literal, LiteralImaginaryUnit, convert_to_literal
@@ -37,12 +38,15 @@ __all__ = (
     'PythonComplex',
     'PythonComplexProperty',
     'PythonConjugate',
+    'PythonDict',
+    'PythonDictFunction',
     'PythonEnumerate',
     'PythonFloat',
     'PythonImag',
     'PythonInt',
     'PythonLen',
     'PythonList',
+    'PythonListFunction',
     'PythonMap',
     'PythonMax',
     'PythonMin',
@@ -722,6 +726,47 @@ class PythonList(TypedAstNode):
         """
         return True
 
+
+class PythonListFunction(PyccelFunction):
+    """
+    Class representing a call to the `list` function.
+
+    Class representing a call to the `list` function. This is
+    different to the `[,]' syntax as it only takes one argument
+    and unpacks any variables.
+
+    Parameters
+    ----------
+    arg : TypedAstNode
+        The argument passed to the function call.
+    """
+    name = 'list'
+    __slots__ = ('_class_type', '_shape')
+    _attribute_nodes = ()
+
+    def __new__(cls, arg):
+        if isinstance(arg, PythonList):
+            return arg
+        elif isinstance(arg.shape[0], LiteralInteger):
+            return PythonList(*[arg[i] for i in range(arg.shape[0])])
+        else:
+            return super().__new__(cls)
+
+    def __init__(self, copied_obj):
+        self._class_type = copied_obj.class_type
+        self._shape = copied_obj.shape
+        super().__init__(copied_obj)
+
+    @property
+    def copied_obj(self):
+        """
+        The object being copied.
+
+        The object being copied to create a new list instance.
+        """
+        return self._args[0]
+
+
 #==============================================================================
 class PythonSet(TypedAstNode):
     """
@@ -810,6 +855,131 @@ class PythonSetFunction(PyccelFunction):
         self._class_type = copied_obj.class_type
         self._shape = copied_obj.shape
         super().__init__(copied_obj)
+
+#==============================================================================
+class PythonDict(PyccelFunction):
+    """
+    Class representing a call to Python's `{}` function.
+
+    Class representing a call to Python's `{}` function which generates a
+    literal Python dict. This operator does not handle `**a` expressions.
+
+    Parameters
+    ----------
+    keys : iterable[TypedAstNode]
+        The keys of the new dictionary.
+    values : iterable[TypedAstNode]
+        The values of the new dictionary.
+    """
+    __slots__ = ('_keys', '_values', '_shape', '_class_type')
+    _attribute_nodes = ('_keys', '_values')
+    _rank = 1
+
+    def __init__(self, keys, values):
+        self._keys = keys
+        self._values = values
+        super().__init__()
+        if pyccel_stage == 'syntactic':
+            return
+        elif len(keys) != len(values):
+            raise TypeError("Unpacking values in a dictionary is not yet supported.")
+        elif len(keys) == 0:
+            self._shape = (LiteralInteger(0),)
+            self._class_type = DictType(GenericType(), GenericType())
+            return
+
+        key0 = keys[0]
+        val0 = values[0]
+        homogeneous_keys = all(k.class_type is not GenericType() for k in keys) and \
+                           all(key0.class_type == k.class_type for k in keys[1:])
+        homogeneous_vals = all(v.class_type is not GenericType() for v in values) and \
+                           all(val0.class_type == v.class_type for v in values[1:])
+
+        if homogeneous_keys and homogeneous_vals:
+            self._class_type = DictType(key0.class_type, val0.class_type)
+
+            self._shape = (LiteralInteger(len(keys)), )
+        else:
+            raise TypeError("Can't create an inhomogeneous dict")
+
+    def __iter__(self):
+        return zip(self._keys, self._values)
+
+    def __str__(self):
+        args = ', '.join(f'{k}: {v}' for k,v in self)
+        return f'{{{args}}}'
+
+    def __repr__(self):
+        args = ', '.join(f'{repr(k)}: {repr(v)}' for k,v in self)
+        return f'PythonDict({args})'
+
+    @property
+    def keys(self):
+        """
+        The keys of the new dictionary.
+
+        The keys of the new dictionary.
+        """
+        return self._keys
+
+    @property
+    def values(self):
+        """
+        The values of the new dictionary.
+
+        The values of the new dictionary.
+        """
+        return self._values
+
+#==============================================================================
+class PythonDictFunction(PyccelFunction):
+    """
+    Class representing a call to the `dict` function.
+
+    Class representing a call to the `dict` function. This is
+    different to the `{}` syntax as it is either a cast function or it
+    uses arguments to create the dictionary. In the case of a cast function
+    an instance of PythonDict is returned to express this concept. In the
+    case of a copy this class stores the description of the copy operator.
+
+    Parameters
+    ----------
+    *args : TypedAstNode
+        The arguments passed to the function call. If args are provided
+        then only one argument should be provided. This object is copied
+        unless it is a temporary PythonDict in which case it is returned
+        directly.
+    **kwargs : dict[TypedAstNode]
+        The keyword arguments passed to the function call. If kwargs are
+        provided then no args should be provided and a PythonDict object
+        will be created.
+    """
+    __slots__ = ('_shape', '_class_type')
+    name = 'dict'
+
+    def __new__(cls, *args, **kwargs):
+        if len(args) == 0:
+            keys = [LiteralString(k) for k in kwargs]
+            values = list(kwargs.values())
+            return PythonDict(keys, values)
+        elif len(args) != 1:
+            raise NotImplementedError("Unrecognised dict calling convention")
+        else:
+            return super().__new__(cls)
+
+    def __init__(self, copied_obj):
+        self._class_type = copied_obj.class_type
+        self._shape = copied_obj.shape
+        super().__init__(copied_obj)
+
+    @property
+    def copied_obj(self):
+        """
+        The object being copied.
+
+        The object being copied to create a new dict instance.
+        """
+        return self._args[0]
 
 #==============================================================================
 class PythonMap(PyccelFunction):
@@ -1254,22 +1424,22 @@ DtypePrecisionToCastFunction = {
 builtin_functions_dict = {
     'abs'      : PythonAbs,
     'bool'     : PythonBool,
-    'range'    : PythonRange,
-    'zip'      : PythonZip,
-    'enumerate': PythonEnumerate,
-    'int'      : PythonInt,
-    'float'    : PythonFloat,
     'complex'  : PythonComplex,
-    'bool'     : PythonBool,
-    'sum'      : PythonSum,
+    'dict'     : PythonDictFunction,
+    'enumerate': PythonEnumerate,
+    'float'    : PythonFloat,
+    'int'      : PythonInt,
     'len'      : PythonLen,
-    'list'     : PythonList,
+    'list'     : PythonListFunction,
+    'map'      : PythonMap,
     'max'      : PythonMax,
     'min'      : PythonMin,
     'not'      : PyccelNot,
-    'map'      : PythonMap,
+    'range'    : PythonRange,
+    'set'      : PythonSetFunction,
     'str'      : LiteralString,
-    'type'     : PythonType,
+    'sum'      : PythonSum,
     'tuple'    : PythonTupleFunction,
-    'set'      : PythonSetFunction
+    'type'     : PythonType,
+    'zip'      : PythonZip,
 }