+++ pyccel/codegen/printing/ccode.py
@@ -11,7 +11,7 @@ from pyccel.ast.basic     import ScopedAstNode
 
 from pyccel.ast.builtins  import PythonRange, PythonComplex
 from pyccel.ast.builtins  import PythonPrint, PythonType
-from pyccel.ast.builtins  import PythonList, PythonTuple, PythonSet
+from pyccel.ast.builtins  import PythonList, PythonTuple, PythonSet, PythonDict
 
 from pyccel.ast.core      import Declare, For, CodeBlock
 from pyccel.ast.core      import FuncAddressDeclare, FunctionCall, FunctionCallArgument
@@ -29,7 +29,7 @@ from pyccel.ast.datatypes import PythonNativeInt, PythonNativeBool, VoidType
 from pyccel.ast.datatypes import TupleType, FixedSizeNumericType
 from pyccel.ast.datatypes import CustomDataType, StringType, HomogeneousTupleType, HomogeneousListType, HomogeneousSetType
 from pyccel.ast.datatypes import PrimitiveBooleanType, PrimitiveIntegerType, PrimitiveFloatingPointType, PrimitiveComplexType
-from pyccel.ast.datatypes import HomogeneousContainerType
+from pyccel.ast.datatypes import HomogeneousContainerType, DictType
 
 from pyccel.ast.internals import Slice, PrecomputedCode, PyccelArrayShapeElement
 
@@ -45,6 +45,8 @@ from pyccel.ast.numpyext import NumpyReal, NumpyImag, NumpyFloat, NumpySize
 from pyccel.ast.numpytypes import NumpyInt8Type, NumpyInt16Type, NumpyInt32Type, NumpyInt64Type
 from pyccel.ast.numpytypes import NumpyFloat32Type, NumpyFloat64Type, NumpyComplex64Type, NumpyComplex128Type
 from pyccel.ast.numpytypes import NumpyNDArrayType, numpy_precision_map
+
+from pyccel.ast.type_annotations import VariableTypeAnnotation
 
 from pyccel.ast.utilities import expand_to_loops
 
@@ -239,6 +241,8 @@ c_imports = {n : Import(n, Module(n, (), ())) for n in
                  'assert',
                  'numpy_c']}
 
+import_header_guard_prefix = {'Set_extensions' : '_TOOLS_SET'}
+
 class CCodePrinter(CodePrinter):
     """
     A printer for printing code in C.
@@ -312,14 +316,22 @@ class CCodePrinter(CodePrinter):
         return self._additional_imports.keys()
 
     def add_import(self, import_obj):
+        """
+        Add a new import to the current context.
+
+        Add a new import to the current context. This allows the import to be recognised
+        at the compiling/linking stage. If the source of the import is not new then any
+        new targets are added to the Import object.
+
+        Parameters
+        ----------
+        import_obj : Import
+            The AST node describing the import.
+        """
         if import_obj.source not in self._additional_imports:
             self._additional_imports[import_obj.source] = import_obj
-
-    def _get_statement(self, codestring):
-        return "%s;\n" % codestring
-
-    def _get_comment(self, text):
-        return "// {0}\n".format(text)
+        elif import_obj.target:
+            self._additional_imports[import_obj.source].define_target(import_obj.target)
 
     def _format_code(self, lines):
         return self.indent_code(lines)
@@ -666,8 +678,13 @@ class CCodePrinter(CodePrinter):
             The generated C code for the container initialization.
         """
 
-        dtype = self.get_c_type(assignment_var.lhs.class_type)
-        keyraw = '{' + ', '.join(self._print(a) for a in expr.args) + '}'
+        class_type = assignment_var.lhs.class_type
+        dtype = self.get_c_type(class_type)
+        if isinstance(expr, PythonDict):
+            dict_item_strs = [(self._print(k), self._print(v)) for k,v in zip(expr.keys, expr.values)]
+            keyraw = '{' + ', '.join(f'{{{k}, {v}}}' for k,v in dict_item_strs) + '}'
+        else:
+            keyraw = '{' + ', '.join(self._print(a) for a in expr.args) + '}'
         container_name = self._print(assignment_var.lhs)
         init = f'{container_name} = c_init({dtype}, {keyraw});\n'
         return init
@@ -996,16 +1013,28 @@ class CCodePrinter(CodePrinter):
             source = source.name[-1]
         else:
             source = self._print(source)
-        if source.startswith('stc/'):
-            stc_name, container_type, container_key = source.split("/")
-            container = container_type.split("_")
-            return '\n'.join((f'#ifndef _{container_type.upper()}',
-                              f'#define _{container_type.upper()}',
-                              f'#define i_type {container_type}',
-                              f'#define i_key {container_key}',
-                              f'#include "{stc_name + "/" + container[0]}.h"',
-                              '#endif\n'))
-
+        if source.startswith('stc/') or source in import_header_guard_prefix:
+            code = ''
+            for t in expr.target:
+                dtype = t.object.class_type
+                container_type = t.target
+                if isinstance(dtype, DictType):
+                    container_key_key = self.get_c_type(dtype.key_type)
+                    container_val_key = self.get_c_type(dtype.value_type)
+                    container_key = f'{container_key_key}_{container_val_key}'
+                    element_decl = f'#define i_key {container_key_key}\n#define i_val {container_val_key}\n'
+                else:
+                    container_key = self.get_c_type(dtype.element_type)
+                    element_decl = f'#define i_key {container_key}\n'
+                header_guard_prefix = import_header_guard_prefix.get(source, '')
+                header_guard = f'{header_guard_prefix}_{container_type.upper()}'
+                code += ''.join((f'#ifndef {header_guard}\n',
+                        f'#define {header_guard}\n',
+                        f'#define i_type {container_type}\n',
+                        element_decl,
+                        f'#include <{source}.h>\n',
+                        f'#endif // {header_guard}\n\n'))
+            return code
         # Get with a default value is not used here as it is
         # slower and on most occasions the import will not be in the
         # dictionary
@@ -1223,10 +1252,17 @@ class CCodePrinter(CodePrinter):
 
             key = (primitive_type, dtype.precision)
         elif isinstance(dtype, (HomogeneousSetType, HomogeneousListType)):
-            container_type = 'hset_' if dtype.name == 'set' else 'vec_'
-            element_type = self.get_c_type(dtype.element_type)
-            i_type = container_type + element_type.replace(' ', '_')
-            self.add_import(Import(f'stc/{i_type}/{element_type}', Module(f'stc/{i_type}', (), ())))
+            container_type = 'hset' if dtype.name == 'set' else 'vec'
+            element_type = self.get_c_type(dtype.element_type).replace(' ', '_')
+            i_type = f'{container_type}_{element_type}'
+            self.add_import(Import(f'stc/{container_type}', AsName(VariableTypeAnnotation(dtype), i_type)))
+            return i_type
+        elif isinstance(dtype, DictType):
+            container_type = 'hmap'
+            key_type = self.get_c_type(dtype.key_type).replace(' ', '_')
+            val_type = self.get_c_type(dtype.value_type).replace(' ', '_')
+            i_type = f'{container_type}_{key_type}_{val_type}'
+            self.add_import(Import(f'stc/{container_type}', AsName(VariableTypeAnnotation(dtype), i_type)))
             return i_type
         else:
             key = dtype
@@ -1303,7 +1339,7 @@ class CCodePrinter(CodePrinter):
         rank  = expr.rank
 
         if rank > 0:
-            if isinstance(expr.class_type, (HomogeneousSetType, HomogeneousListType)):
+            if isinstance(expr.class_type, (HomogeneousSetType, HomogeneousListType, DictType)):
                 dtype = self.get_c_type(expr.class_type)
                 return dtype
             if isinstance(expr.class_type,(HomogeneousTupleType, NumpyNDArrayType)):
@@ -1436,16 +1472,9 @@ class CCodePrinter(CodePrinter):
         inds = list(expr.indices)
         base_shape = base.shape
         allow_negative_indexes = expr.allows_negative_indexes
-        if isinstance(base.class_type, NumpyNDArrayType):
-            #set dtype to the C struct types
-            dtype = self.find_in_ndarray_type_registry(expr.dtype)
-        elif isinstance(base.class_type, HomogeneousContainerType):
-            dtype = self.find_in_ndarray_type_registry(numpy_precision_map[(expr.dtype.primitive_type, expr.dtype.precision)])
-        else:
-            raise NotImplementedError(f"Don't know how to index {expr.class_type} type")
 
         if isinstance(base, IndexedElement):
-            while isinstance(base, IndexedElement) and isinstance(base.class_type, HomogeneousContainerType):
+            while isinstance(base, IndexedElement) and isinstance(base.class_type, (NumpyNDArrayType, HomogeneousTupleType)):
                 inds = list(base.indices) + inds
                 base = base.base
 
@@ -1461,7 +1490,27 @@ class CCodePrinter(CodePrinter):
                     inds[i] = IfTernaryOperator(PyccelLt(ind, LiteralInteger(0)),
                         PyccelAdd(base_shape[i], ind, simplify = True), ind)
 
+        if isinstance(base.class_type, HomogeneousListType):
+            assign = expr.get_user_nodes(Assign)
+            index = self._print(inds[0])
+            list_var = self._print(ObjectAddress(base))
+            container_type = self.get_c_type(base.class_type)
+            if assign:
+                assert len(assign) == 1
+                assign_node = assign[0]
+                lhs = assign_node.lhs
+                if lhs == expr or lhs.is_user_of(expr):
+                    return f"(*{container_type}_at_mut({list_var},{index}))"
+            return f"(*{container_type}_at({list_var},{index}))"
+
         base_name = self._print(base)
+        if isinstance(base.class_type, NumpyNDArrayType):
+            #set dtype to the C struct types
+            dtype = self.find_in_ndarray_type_registry(expr.dtype)
+        elif isinstance(base.class_type, HomogeneousTupleType):
+            dtype = self.find_in_ndarray_type_registry(numpy_precision_map[(expr.dtype.primitive_type, expr.dtype.precision)])
+        else:
+            raise NotImplementedError(f"Don't know how to index {expr.class_type} type")
         if expr.rank > 0:
             #managing the Slice input
             for i , ind in enumerate(inds):
@@ -1598,7 +1647,7 @@ class CCodePrinter(CodePrinter):
     def _print_Allocate(self, expr):
         free_code = ''
         variable = expr.variable
-        if isinstance(variable.class_type, (HomogeneousListType, HomogeneousSetType)):
+        if isinstance(variable.class_type, (HomogeneousListType, HomogeneousSetType, DictType)):
             return ''
         if variable.rank > 0:
             #free the array if its already allocated and checking if its not null if the status is unknown
@@ -1634,7 +1683,7 @@ class CCodePrinter(CodePrinter):
             raise NotImplementedError(f"Allocate not implemented for {variable}")
 
     def _print_Deallocate(self, expr):
-        if isinstance(expr.variable.class_type, (HomogeneousListType, HomogeneousSetType)):
+        if isinstance(expr.variable.class_type, (HomogeneousListType, HomogeneousSetType, DictType)):
             variable_address = self._print(ObjectAddress(expr.variable))
             container_type = self.get_c_type(expr.variable.class_type)
             return f'{container_type}_drop({variable_address});\n'
@@ -2184,10 +2233,17 @@ class CCodePrinter(CodePrinter):
         if isinstance(rhs, (NumpyFull)):
             return prefix_code+self.arrayFill(expr)
         lhs = self._print(expr.lhs)
-        if isinstance(rhs, (PythonList, PythonSet)):
+        if isinstance(rhs, (PythonList, PythonSet, PythonDict)):
             return prefix_code+self.init_stc_container(rhs, expr)
         rhs = self._print(expr.rhs)
         return prefix_code+'{} = {};\n'.format(lhs, rhs)
+
+    def _print_SetPop(self, expr):
+        dtype = expr.set_variable.class_type
+        var_type = self.get_c_type(dtype)
+        self.add_import(Import('Set_extensions', AsName(VariableTypeAnnotation(dtype), var_type)))
+        set_var = self._print(ObjectAddress(expr.set_variable))
+        return f'{var_type}_pop({set_var})'
 
     def _print_AliasAssign(self, expr):
         lhs_var = expr.lhs