+++ tests/errors/semantic/blocking/TYPE_ALIAS_REASSIGN.py
@@ -1,2 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
-a = {'a': 1, 'b':2}
+from typing import TypeAlias
+
+MyType : TypeAlias = int
+
+MyType = float